+++ test/erlang_package/pinned_deb/default_spec.rb
@@ -9,7 +9,7 @@
 ).each do |p|
   describe package(p) do
     it { should be_installed }
-    its('version') { should eq '1:21.2.5-1' }
+    its('version') { should eq '1:21.2.7-1' }
   end
 end